+++ src/dsp/lpf.h
@@ -7,7 +7,7 @@
 
 typedef struct lpf_t lpf;
 
-int lpf_create(uint8_t decimation, uint32_t sampling_freq, uint32_t cutoff_freq, uint32_t transition_width, size_t output_len, size_t num_bytes, lpf **filter);
+int lpf_create(uint8_t decimation, uint64_t sampling_freq, uint64_t cutoff_freq, uint32_t transition_width, size_t output_len, size_t num_bytes, lpf **filter);
 
 void lpf_process(const void *input, size_t input_len, void **output, size_t *output_len, lpf *filter);
